refactoring tidyup
