+++ src/main/scala/actors/messages/DoUpdate.scala
@@ -1,0 +1,5 @@
+package actors.messages
+
+import com.gu.contentatom.thrift.Atom
+
+case class DoUpdate(atom:Atom)
