+++ src/main/scala/actors/messages/ErrorSend.scala
@@ -1,0 +1,5 @@
+package actors.messages
+
+case class ErrorSend(error:String) extends Throwable {
+  override def getMessage: String = error
+}
