+++ src/main/scala/actors/messages/GotPlutoId.scala
@@ -1,0 +1,3 @@
+package actors.messages
+
+case class GotPlutoId(plutoId:String)
