+++ src/main/scala/actors/messages/LookupPlutoId.scala
@@ -1,0 +1,3 @@
+package actors.messages
+
+case class LookupPlutoId(atomId:String)
