+++ src/main/scala/actors/messages/NoResponseBody.scala
@@ -1,0 +1,3 @@
+package actors.messages
+
+case class NoResponseBody()
