+++ src/main/scala/actors/messages/SuccessfulSend.scala
@@ -1,0 +1,3 @@
+package actors.messages
+
+case class SuccessfulSend()
