+++ src/main/scala/vidispine/VsXml.scala
@@ -1,4 +1,7 @@
+package vidispine
+
 import scala.xml.{Node, NodeSeq}
+
 trait VsXml {
   def valuesForField(xmlNode:NodeSeq, fieldName: String):Seq[String] = {
     val field = (xmlNode \ "timespan" \ "field").filter({node:Node=> (node \ "name").text==fieldName})
